App-Running/requirements.txt
torchvision==0.10.0
numpy==1.18.5
opencv_python==4.6.0.66
torch==1.9.0
albumentations==1.1.0
matplotlib==3.4.3
mediapipe==0.8.7.1
pytorch_lightning==1.4.9
PyYAML==6.0